requirements.txt
ultralytics
roboflow
torch
torchvision
torchaudio
numpy
opencv-python
matplotlib
tqdm
pillow
seaborn